+++ ClassLibrary1/Protocols/Protocol_EP06.cs
@@ -12,20 +12,20 @@ namespace CentralLib.Protocols
 
         //UInt16 MaxStringLenght = 75;
 
-        public Protocol_EP06(int serialPort, int inFPNumber):base(serialPort, inFPNumber)
+        public Protocol_EP06(int serialPort, Int64 inFPNumber):base(serialPort, inFPNumber)
         {
             MaxStringLenght = 75;
             useCRC16 = false;
             //initial();
         }
 
-        public Protocol_EP06(CentralLib.Connections.DefaultPortCom dComPort,int inFPNumber) :base(dComPort, inFPNumber)
+        public Protocol_EP06(CentralLib.Connections.DefaultPortCom dComPort,Int64 inFPNumber) :base(dComPort, inFPNumber)
         {
             MaxStringLenght = 75;
             useCRC16 = false;
         }
 
-        public Protocol_EP06(string IpAdress, int port, int inFPNumber):base(IpAdress,port, inFPNumber)
+        public Protocol_EP06(string IpAdress, int port, Int64 inFPNumber):base(IpAdress,port, inFPNumber)
         {
             MaxStringLenght = 75;
             useCRC16 = false;
