+++ PrintFP/Program.cs
@@ -17,7 +17,7 @@ namespace PrintFP
     public static class Program 
     {
         private static string fpnumber, server;
-        private static int FPnumber;
+        private static long FPnumber;
         private static Logger logger = LogManager.GetCurrentClassLogger();
         private static DateTime startJob;
         //private static System.Timers.Timer _timer;
@@ -48,7 +48,7 @@ namespace PrintFP
             try
             {
                 var p = os.Parse(args);
-                FPnumber = int.Parse(fpnumber);
+                FPnumber = long.Parse(fpnumber);
             }
             catch (Exception e)
             {
